GTNPORTAL-2657 Gatein master Unit tests fails in GateIn Portal Component Portal Data on oracle7
This failure happens if testClosureCompressing is executed before testPriority. testClosureCompressing registers a plugin with the ResourceCompressorService, but doesn't unregister it after the test has completed.
As the ResourceCompressorService doesn't have an unregister method I've added a unregister method to the test that uses reflection to remove it from ResourceCompressorService.plugins. An alternative fix would be to add a unregister method to the ResourceCompressorService and use that instead.