GTNPORTAL-3359 - Supressed WSRP-related exceptions on shutdown
Supressed the exception thrown in case something goes wrong while shutting down the consumers, as we cannot guarantee that the stop routine for WSRP will be called before the stop routine for the JCR connection manager. The proper solution would be to find a way to hook into the JCR shutdown and stop it before the JCR connection manager shuts down, but there doesn't seems to be a way to do it.